Rename crust_mock to mock_crust for consistency
